move training function to train script, replace log_dir with run_dir
